Move function sam_to_fastq from clusters to mapping module
