+++ ariba/tests/data/mapping_test_sam_to_fastq.bam
